Finalizing API v6 for translated values of the handler. Code review corrections.

+++ bundles/org.openhab.binding.pihole/src/main/java/org/openhab/binding/pihole/internal/rest/model/v6/StatAnswer.java
@@ -25,7 +25,7 @@ public record StatAnswer(Queries queries, Clients client, Gravity gravity, doubl
 
     }
 
-    public record Gravity(int domainsBeingBlocked, int lastUpdate) {
+    public record Gravity(int domainsBeingBlocked, long lastUpdate) {
 
     }
